Fix data types and formatting in temperature conversion program

+++ lesson10.c
@@ -4,12 +4,12 @@
 #define STEP 20
 int main()
 {
-    int fahr, celsius;
+    float fahr, celsius;
     fahr = INTI_VALUE;
     while (fahr <= MAX_VALUE)
     {
-        celsius = (5 * (fahr - 32) / 9);
-        printf("%3d %3d\n", fahr, celsius);
+        celsius = (5.0 * (fahr - 32.0) / 9.0);
+        printf("%3.0f %6.2f\n", fahr, celsius);
         fahr += STEP;
     }
     return 0;
